// frontend/src/types/csaf.ts
export type DocumentAcknowledgments = [Acknowledgment, ...Acknowledgment[]];
// ...
export type DocumentCategory = string;
/**
 * Gives the version of the CSAF specification which the document was generated for.
 */
export type CSAFVersion = "2.0";
// ...
export type DocumentLanguage = string;
/**
 * Holds notes associated with the whole document.
 */
export type DocumentNotes = [Note, ...Note[]];
/**
 * Indicates who is intended to read it.
 */
export type AudienceOfNote = string;
/**
 * Contains the information of what kind of note this is.
 */
export type NoteCategory =
  | "description"
  | "details"
  | "faq"
  | "general"
  | "legal_disclaimer"
  | "other"
  | "summary";
/**
 * Holds the content of the note. Content varies depending on type.
 */
export type NoteContent = string;
/**
 * Provides a concise description of what is contained in the text of the note.
 */
export type TitleOfNote = string;
// ...
export type DocumentReferences = [Reference, ...Reference[]];
/**
 * Indicates whether the reference points to the same document or vulnerability in focus (depending on scope) or to an external resource.
 */
export type CategoryOfReference = "external" | "self";
/**
 * Indicates what this reference refers to.
 */
export type SummaryOfTheReference = string;
/**
 * Provides the URL for the reference.
 */
export type URLOfReference = string;
/**
 * If this copy of the document is a translation then the value of this property describes from which language this document was translated.
 */
export type SourceLanguage = string;
/**
 * This SHOULD be a canonical name for the document, and sufficiently unique to distinguish it from similar documents.
 */
export type TitleOfThisDocument = string;
// ...
export type ListOfBranches = [Branch, ...Branch[]];
/**
 * Describes the characteristics of the labeled branch.
 */
export type CategoryOfTheBranch =
  | "architecture"
  | "host_name"
  | "language"
  | "legacy"
  | "patch_level"
  | "product_family"
  | "product_name"
  | "product_version"
  | "product_version_range"
  | "service_pack"
  | "specification"
  | "vendor";
/**
 * Contains the canonical descriptor or 'friendly name' of the branch.
 */
export type NameOfTheBranch = string;
/**
 * The value should be the product’s full canonical name, including version number and other attributes, as it would be used in a human-friendly document.
 */
export type TextualDescriptionOfTheProduct = string;
/**
 * Token required to identify a full_product_name so that it can be referred to from other parts in the document. There is no predefined or required format for the product_id as long as it uniquely identifies a product in the context of the current document.
 */
export type ReferenceTokenForProductInstance = string;
// ...
export type ListOfFullProductNames = [FullProductName, ...FullProductName[]];
/**
 * Contains a list of product groups.
 */
export type ListOfProductGroups = [ProductGroup, ...ProductGroup[]];
// ...
export type ListOfRelationships = [Relationship, ...Relationship[]];
// ...
export type Vulnerabilities = [Vulnerability, ...Vulnerability[]];
// ...
export interface CommonSecurityAdvisoryFramework {
  document: DocumentLevelMetaData;
  product_tree?: ProductTree;
  vulnerabilities?: Vulnerabilities;
  [k: string]: unknown;
}
/**
 * Captures the meta-data about this document describing a particular set of security advisories.
 */
export interface DocumentLevelMetaData {
  acknowledgments?: DocumentAcknowledgments;
  aggregate_severity?: AggregateSeverity;
  category: DocumentCategory;
  csaf_version: CSAFVersion;
  distribution?: RulesForSharingDocument;
  lang?: DocumentLanguage;
  notes?: DocumentNotes;
  publisher: Publisher;
  references?: DocumentReferences;
  source_lang?: SourceLanguage;
  title: TitleOfThisDocument;
  tracking: Tracking;
  [k: string]: unknown;
}
// ...
export interface AggregateSeverity {
  namespace?: NamespaceOfAggregateSeverity;
  text: TextOfAggregateSeverity;
  [k: string]: unknown;
}
/**
 * Describe any constraints on how this document might be shared.
 */
export interface RulesForSharingDocument {
  text?: TextualDescription;
  tlp?: TrafficLightProtocolTLP;
  [k: string]: unknown;
}
/**
 * Provides details about the TLP classification of the document.
 */
export interface TrafficLightProtocolTLP {
  label: LabelOfTLP;
  url?: URLOfTLPVersion;
  [k: string]: unknown;
}
/**
 * Is a place to put all manner of text blobs related to the current context.
 */
export interface Note {
  audience?: AudienceOfNote;
  category: NoteCategory;
  text: NoteContent;
  title?: TitleOfNote;
  [k: string]: unknown;
}
/**
 * Provides information about the publisher of the document.
 */
export interface Publisher {
  category: CategoryOfPublisher;
  contact_details?: ContactDetails;
  issuing_authority?: IssuingAuthority;
  name: NameOfPublisher;
  namespace: NamespaceOfPublisher;
  [k: string]: unknown;
}
/**
 * Holds any reference to conferences, papers, advisories, and other resources that are related and considered related to either a surrounding part of or the entire document and to be of value to the document consumer.
 */
export interface Reference {
  category?: CategoryOfReference;
  summary: SummaryOfTheReference;
  url: URLOfReference;
  [k: string]: unknown;
}
/**
 * Is a container designated to hold all management attributes necessary to track a CSAF document as a whole.
 */
export interface Tracking {
  aliases?: Aliases;
  current_release_date: CurrentReleaseDate;
  generator?: DocumentGenerator;
  id: UniqueIdentifierForTheDocument;
  initial_release_date: InitialReleaseDate;
  revision_history: RevisionHistory;
  status: DocumentStatus;
  version: Version;
  [k: string]: unknown;
}
// ...
export interface ProductTree {
  branches?: ListOfBranches;
  full_product_names?: ListOfFullProductNames;
  product_groups?: ListOfProductGroups;
  relationships?: ListOfRelationships;
  [k: string]: unknown;
}
/**
 * Is a part of the hierarchical structure of the product tree.
 */
export interface Branch {
  branches?: ListOfBranches;
  category: CategoryOfTheBranch;
  name: NameOfTheBranch;
  product?: FullProductName;
  [k: string]: unknown;
}
/**
 * Specifies information about the product and assigns the product_id.
 */
export interface FullProductName {
  name: TextualDescriptionOfTheProduct;
  product_id: ReferenceTokenForProductInstance;
  product_identification_helper?: HelperToIdentifyTheProduct;
  [k: string]: unknown;
}
/**
 * Provides at least one method which aids in identifying the product in an asset database.
 */
export interface HelperToIdentifyTheProduct {
  cpe?: CommonPlatformEnumerationRepresentation;
  hashes?: ListOfHashes;
  model_numbers?: ListOfModels;
  purl?: PackageURLRepresentation;
  sbom_urls?: ListOfSBOMURLs;
  serial_numbers?: ListOfSerialNumbers;
  skus?: ListOfStockKeepingUnits;
  x_generic_uris?: ListOfGenericURIs;
  [k: string]: unknown;
}
// ...
export interface Vulnerability {
  acknowledgments?: VulnerabilityAcknowledgments;
  cve?: CVE;
  cwe?: CWE;
  discovery_date?: DiscoveryDate;
  flags?: ListOfFlags;
  ids?: ListOfIDs;
  involvements?: ListOfInvolvements;
  notes?: VulnerabilityNotes;
  product_status?: ProductStatus;
  references?: VulnerabilityReferences;
  release_date?: ReleaseDate;
  remediations?: ListOfRemediations;
  scores?: ListOfScores;
  threats?: ListOfThreats;
  title?: Title;
  [k: string]: unknown;
}